+++ CHANGES.rst
@@ -15,6 +15,7 @@ Imviz
 
 Mosviz
 ^^^^^^
+- NIRISS parser now sorts FITS files by header instead of file name. [#819]
 
 Specviz
 ^^^^^^^
@@ -69,7 +70,6 @@ Cubeviz
 
 Mosviz
 ^^^^^^
-
 - Added ``--instrument`` CLI option to support NIRISS data loading in Mosviz. [#1488]
 
 Bug Fixes